# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FB Agency Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      security:
        - bearerAuth: []
      summary: List all managed pages
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, all]
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacebookPage"
    post:
      operationId: createPage
      tags: [pages]
      security:
        - bearerAuth: []
      summary: Add a page to management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageBody"
      responses:
        "201":
          description: Page added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacebookPage"
...
  /pages/{pageId}/automation:
    patch:
      operationId: updatePageAutomation
      tags: [pages]
      security:
        - bearerAuth: []
      summary: Update page automation settings
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageAutomationBody"
      responses:
        "200":
          description: Updated page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacebookPage"
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      security:
        - bearerAuth: []
      summary: List token packages
      responses:
        "200":
          description: Token balance and packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokensInfo"
    post:
      operationId: addTokens
      tags: [tokens]
      security:
        - bearerAuth: []
      summary: Add tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTokensBody"
      responses:
        "200":
          description: Updated token balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokensInfo"
...
  /youtube-accounts:
    get:
      operationId: listYoutubeAccounts
      tags: [youtube-accounts]
      security:
        - bearerAuth: []
      summary: List connected Google/YouTube accounts
      responses:
        "200":
          description: List of YouTube accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YouTubeAccount"
  /youtube-accounts/{accountId}:
    delete:
      operationId: deleteYoutubeAccount
      tags: [youtube-accounts]
      security:
        - bearerAuth: []
      summary: Disconnect a Google/YouTube account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Account disconnected
  /youtube-accounts/{accountId}/sync-channels:
    post:
      operationId: syncYoutubeAccountChannels
      tags: [youtube-accounts]
      security:
        - bearerAuth: []
      summary: Re-sync channels from YouTube API for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channels synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncYoutubeChannelsResponse"
...
  /automation-logs:
    get:
      operationId: listAutomationLogs
      tags: [logs]
      security:
        - bearerAuth: []
      summary: List automation logs
      parameters:
        - name: pageId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of automation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationLog"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FacebookPage:
      type: object
      properties:
        id:
          type: string
        fbPageId:
          type: string
        name:
          type: string
        category:
          type: string
        profilePicture:
          type: string
        followersCount:
          type: integer
        automationEnabled:
          type: boolean
        postingFrequency:
          type: string
          enum: [daily, twice_daily, weekly, custom]
        status:
          type: string
          enum: [active, paused, error]
        accountId:
          type: string
        lastPostedAt:
          type: string
        sourceType:
          type: string
          enum: [instagram, youtube, tiktok]
        sourceIdentity:
          type: string
        postsPerDay:
          type: integer
        scheduleLogic:
          type: string
          enum: [fixed, random]
        timezone:
          type: string
        timeSlots:
          type: array
          items:
            type: string
        scrapingStatus:
          type: string
          enum: [pending, active, done]
        totalPosted:
          type: integer
        totalPending:
          type: integer
        totalFailed:
          type: integer
        createdAt:
          type: string
      required: [id, fbPageId, name, automationEnabled, status, accountId, createdAt, postsPerDay, scheduleLogic, timezone, scrapingStatus, totalPosted, totalPending, totalFailed]
    CreatePageBody:
      type: object
      properties:
        fbPageId:
          type: string
        name:
          type: string
        category:
          type: string
        accountId:
          type: string
        postingFrequency:
          type: string
          enum: [daily, twice_daily, weekly, custom]
        sourceType:
          type: string
          enum: [instagram, youtube, tiktok]
        sourceIdentity:
          type: string
        postsPerDay:
          type: integer
        scheduleLogic:
          type: string
          enum: [fixed, random]
        timezone:
          type: string
      required: [fbPageId, name, accountId]
...
    UpdatePageAutomationBody:
      type: object
      properties:
        postsPerDay:
          type: integer
        scheduleLogic:
          type: string
          enum: [fixed, random]
        timezone:
          type: string
        timeSlots:
          type: array
          items:
            type: string
        automationEnabled:
          type: boolean
...
    TokensInfo:
      type: object
      properties:
        balance:
          type: integer
        packages:
          type: array
          items:
            $ref: "#/components/schemas/TokenPackage"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TokenTransaction"
      required: [balance, packages, transactions]
    TokenPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tokens:
          type: integer
        price:
          type: number
      required: [id, name, tokens, price]
    TokenTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [purchase, usage, refund]
        amount:
          type: integer
        description:
          type: string
        timestamp:
          type: string
      required: [id, type, amount, description, timestamp]
    AddTokensBody:
      type: object
      properties:
        packageId:
          type: string
        amount:
          type: integer
      required: [amount]
    AutomationLog:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        message:
          type: string
        pageId:
          type: string
        pageName:
          type: string
        accountId:
          type: string
        status:
          type: string
          enum: [success, error, info]
        metadata:
          type: string
        createdAt:
          type: string
      required: [id, type, message, status, createdAt]
...
    YouTubeAccount:
      type: object
      properties:
        id:
          type: string
        googleId:
          type: string
        name:
          type: string
        email:
          type: string
        profilePicture:
          type: string
        channelsCount:
          type: integer
        status:
          type: string
          enum: [connected, expired, error]
        connectedAt:
          type: string
      required: [id, googleId, name, channelsCount, status, connectedAt]
    SyncYoutubeChannelsResponse:
      type: object
      properties:
        synced:
          type: integer
      required: [synced]
...